system: rewrite trust integration for certctl use
Since OPNsense 22.1 we are using FreeBSD 13 and it comes with a
base trust store which is also maintained there.  In order to be
user-configurable there is also a tool called certctl which will
manage blocking and filling the OpenSSL trust store location
/etc/ssl/certs.  The idea is to make this implicit and faster.

This, however, pseudo-obsoletes the trust bundle handling which
we mainly operate through /etc/ssl/cert.pem.  By pseudo I mean
that ports will still want the real bundles and/or know/guess
this location at complile time.  curl has such overrides for
example.

ca_root_nss's bundle is also pulled in thorough certctl so we
are going to have to jump through a few hoops now in order to
add our certificates cleanly and "prevent" breakage of the
resulting trust store.

Therefore now we write our CA content into separate files because
certctl only hashes the first certificate found in the file.
This is already a bit problematic for ca_root_nss having a
larage number of files in it...  And against all odds the
first certificate I wrote for our bundle is blacklisted by
FreeBSD which made certctl discard all OPNsense authorities
added from the GUI.

To avoid further issues with certclt as a broker here I have
added it in passthru() mode to see eventual errors clearly.

Now when certcl is done all the files are linked in the
/etc/ssl/certs directory but we actually have to build the
full bundle for compatibility with old ports requiring one
of the locations that ca_root_nss ETCSYMLINK option provides.

A shortcoming of certctl is the lack of a bundle mode for
compatibility's sake which is causing a number of problems in
the ports tree at the moment (which is why we do this work now
and take a closer look before this is rolled out in full in
FreeBSD ports).

The bundle is created by iterating over all files in /etc/ssl/certs
and putting them in the expected locations.  One caveat is that
this bloats the bundles to 1.5MB from previously 750KB.  The whole
process is a lot slower, especially certctl doing the rehash.

Long story short: this is going to cause issues in the long run,
but for now we know how it is supposed to work and are ready
for FreeBSD ports to drop support for bundles in individual ports.
But that being said we will probably drag the bundles on for
a few years anyway.