system: adjust the dialog presentation a bit
The danger representation is a bit over the top for deleting
a backup.  The revert is a bit too lax.  So move both to a
warning and indicate that the yes button is the one that should
be clicked in these cases, but in the same warning color as the
modal itself.  While here change the title to something more
appropriate.