firmware: fix slowness in applying firmware settings
We don't even need the full rc.configure_firmware script as that
is for after a core package was updated.  Rather we just want the
actual firmware settings to reload so we add a different path for
it.  Now it is faster than it ever was.

(cherry picked from commit 96b47255be612ae010f5f1caddbac5844e1c4916)