Kea / Dhcp - DHCPv4 replacement [https://github.com/opnsense/core/issues/6971], add "Auto collect option data" checkbox on the subnet.
Our current default (legacy) scenario is to use the configured interface addresss as gateway and primary dns, but since we don't want to complicate the templates and make the choice explicit, we're adding an option to auto-collect and persist on configuration save/apply.

This commit auto updates selected option data fields on request with the first matching address for the requested subnet found in ifconfig.