system: prevent activating shell for non-admins
Our policy has always been that shell access should not
not be granted to untrusted users.  Apparently this is
also a way to wedge in security reports stemming from
this misconfiguration while glossing over this point and
focusing on something else found in the file system.

If you need access for a user just give them full admin
privileges and that's fixed. For everyone else no longer
render the shell and add a warning in the user edit page.