dhcp: do not clobber $range_to / $range_from with the legacy test merge #7046
The merge is used as a test if a prefix is set at all (the legacy input
required a "prefix" but it was actually a suffix and verified as such)
but now that we prevent merging without a leading "::" the final compress
moves the compressed format from the front to the end because that sequence
is longer but the next merge doesn't like that.  Do the test merge without
storing the result as we do not need it anyway.