System: Settings: Tunables - change zfs transcation group defaults. (timeout from 5 --> 90 seconds, dirty data sync percentage from 20 --> 5)
Usually firewalls are not processing a lot of disk write operations, in which case a forced every 5 second flush will lead to quite some wear on the underlying storage device (usually a solid state one these days).
Roughly 20GB per day is being written when using the upstream defaults and without much actual load. To compensate a bit for the reduction of forced syncs, we lower the threshold on the dirty data.