wireguard: improve previous
wg_start() can detect if the interface was removed, which is
ensured during 'restart' and a fresh 'configure'.  The device
could have been created from wireguard_prepare() in the plugin
code but that should normally be used by interfaces_configure()
which also configures the interface correctly (same as the
interfaces_restart_by_device() call).

We only reload the routing in such cases now either as in the
other case the routes should have been placed and remain.