mvc - extend model implementation to ease legacy migrations.
The complexity of some legacy modules such as the gateways indicate that it would be practical to offer something in between legacy and mvc. Normal modals require a single container to account for versions and easy access, where some legacy objects miss the container level in which case overlaying is not possible.

Without the singular container level, it's not possible to track versions and offfer migrations, but does prevent data being moved to different spots in the config.xml and break existing users of the same data.

This commit adds the isLegacyMapper() method, which is controlled by the model mountpoint using an additional plus (+) sign at the end of the mount. e.g. /cert+ would collect all items the xpath /opnsense/cert would offer and hooks them on items with the same array key.

For example, the model below would offer access to all cert items (and thus add's an uuid to them):
<model>
    <mount>/cert+</mount>
    <version>1.0.0</version>
    <items>
        <cert type="ArrayField">
            <refid type="TextField"/>
            <descr type="DescriptionField"/>
            <caref type="TextField"/>
            <crt type="TextField"/>
            <prv type="TextField"/>
        </cert>
    </items>
</model>

In the long run this would help migrate these kind of objects with minimal risk of breakage, making a full model migration at a later moment in time easier (which requires data to be moved into a single container).

Requirement for https://github.com/opnsense/core/issues/7248