interfaces: convert GIF configuration to MVC/API , closes https://github.com/opnsense/core/issues/7241
minor modification in reconfigure_gifs.php to prevent ipv6 addresses to be left behind after a change.
When the interface already exists, flush all addresses in the same way Wireguard is doing on our end and link virtual ips and routes back after configuration.

For IPv4 this seems to be less problematic as additional addresses are aliases and changing the address leads to an update instead of an "insert".