Config - fix locking issue when already owning the lock.
In most cases we use config locking, the lock is acquired only once, in which case all works as expected.
When a controller fetches a lock and passes it on to one of the core controller wrappers which then re-acquires the same lock, the default is to load the config again.
Pending data in other models will be flushed in these cases, which is unexpected if we're updating the same config, but a different section.

This change only executes the reload when not yet locked by this process.