System: Access: Users - change user certificate handling to align with the new MVC module.
Remove the previous link from the user to the certificate as this suggest some relation between the two that doesn't exist (other than the name).
In the userview, the certificate list contains all client certs with the same commonname and verifies if the logged in user is able to access the trust module to prevent data leakage.

closes https://github.com/opnsense/core/issues/7387