interfaces: evidence suggests + 2 is better than + 1 #7400
The + 1 was completely arbitrary to begin with (derived from
FreeBSD scripting), but if part of the system needs longer to
cope with tentative state then this would be an easy way to
make it more reliable.

If + 3 makes sense for the next person is something I want to
doubt, however.

Special thanks go to @Wireheadbe for pursuing and testing this.

(cherry picked from commit f2e60c18f66a4bcf712078ea718fdb6a95385f55)