interfaces: present interface ID in hex
We have to revoke being able to set "random" here for now, but
it's not a huge issue since the fallback is now in the daemon.
The daemon also supports the full unsigned long long range but
does not allow hex input which means we are stuck with losing the
highest bit for now if we store in integer in PHP.

We could store a string, but then cannot verify it and it would
not be a hex value.  Meh.