dashboard: simplify error handling and introduce retry mechanism for timed out requests
Due to the volume of requests being fired, either the browser or server will introduce
some limitation somewhere that cannot easily be trapped, as a result, widgets
were being loaded inconsistently.
The way forward is to use the $.ajax timeout feature that allows requests
to be cancelled if queued up indefinitely by the browser. Due to the way this
generic new ajaxGet function is wrapped by a Promise, we don't have to add
error handling unless we wish to override it. The widget manager will take
care of failed promises both during load and during the tick routine.