webgui: long lasting calls and locking the request pipeline. work-around for https://github.com/opnsense/core/issues/7651
The number of php processes we offer to lighttpd is quite low (2 x 3), which results in a global lock when 2 dashboards are open at the same time.
Although we should increase the limit, we also should figure out a way to limit the number of long running processes as well.

(cherry picked from commit 724f8494d64074e0384b9f593e699daf2d84d604)