dashboard: monit widget: Invert the logic of success and failed
When a status code is not mapped, it will always fall back to failed in the tooltip,
and text-danger as color. Fixes issue with ping check, the status is 16384 when
failed. Since it can be assumed that all statuses other than 0 are bad, this change
in logic seems pragmatic. (#7705)

(cherry picked from commit 8ed829be0597b86175255c11a72fe344e7977629)