interfaces: transform this a bit more #7446
Ideally we should really handle an array of parent hardware
devices here.  In some cases IPv6 device is missed as well.
And later on the code will not check if this was a device
node for a modem so wrapping the hardware stuff into a function
for iteration is the way to go.