interfaces: lock down PPP modes #7446
This change reflects what PPP is actually capable off under
the hood and no functional changes intended.

You could still post a different type but since the goal is
to get rid of posting anything related to PPP this will make
it easier to judge if people actually used this.