firmware: move cleanups to opnsense-update "-F"
This probably needs more engineering later on which makes
it a little harder to ship but in reality we want people
to run the portable command easily in order to unbreak their
environments as pkg-upgrade tripping over its own temporary
files and directories keeps happening in user reports.

While here use "-s" while at it to speed up the script call.
This also pertains to "-e" which is its sibling.