System: Trust: Certificates - link certificates by subject on import, closes https://github.com/opnsense/core/issues/7813
It looks like we only linked certificates on CA import, move that code into the Cert model and improve it a bit, also trigger when importing a new certificate manually now.
To improve the matching, we search for subjects that match best based on item length now (e.g. a dn ZH,NL should precede one with only NL)