dashboard: fix ACL check on dashboard restore and add safety check for save action
Parsing the persisted dashboard requires decoding it, which by extension means we don't need to encode it anymore when returning the dashboard. In all cases, this commit now makes sure that dashboard is filtered according to access rights.

While here, invalid input can be filtered out partly by check the result of json_decode(), so do this when saving the widgets.