interfaces: fix issue with PPP port not being posted; closes #7823
Unify the situation for 'ppp_port' vs 'ports' use by using the latter
as is.  One of the issues of this page always was inconsistency after
all.  Then also zap the validation that causes the changes not to be
accepted.  In theory it's posted anyway but that has been the case for
PPPoE/L2TP/PPTP since the fork.