interfaces: move IPv6 connectivity to a separate script for #7446
A number of cleanup and non-invasive restructuring efforts around
the way IPv6-based connectivity is handled in PPP environments so
that we can later use it allow IPv6-only deployments which takes
a different path based on whether PPP is allowed to do IPv4 (with
or without IPv6) or IPv6 without IPv4.

In theory this would also allow us to run DHCP over PPP, but it's
much outside the scope of the current effort and certainly needs
someone with that type of setup...  ;)

While here make sure the IPv4 connectivity abuse in DHCP scenarios
still somewhat works at least for 24.7.x.  We don't want it unless
we have a strict need for it or may want to come up with a simpler
approach.