system: render header for failed active widgets to allow removal (#7858)
* system: render header for failed active widgets to allow removal

If there is a syntax error in the widget JS code we will not be
able to remove it from the dashboard.  This is only a POC as it
doesn't fully work for varying reasons, but details where the
problem lies.

* dashboard: account for failed module imports or class instantiations

---------

Co-authored-by: Stephan de Wit <stephan.de.wit@deciso.com>
(cherry picked from commit c31f22f23ff40e408d43e5dc3d2a99ba9cf55765)