plugins: add plugins_argument_map() helper
This helper parses "mapped" arguments for batch reloads of e.g.
interfaces or devices or items related to that (like gateway
names).  The rules are simple:

o If the mapped argument is null reload in full (return true)
o If the mapped argument is an array handle it under later conditionals (return true)
o If the mapped argument is of an unsupported type do nothing (return false)
o If the mapped argument is not null but empty do nothing (return false)
o If the argument is a string convert it to an array, splitting the string
  by comma (return true)
o Modify the $map argument to be either null or a non-empty array to normalize
  the passed map (also for backwards-compatible reasons of newwanip_map
  and vpn_map intermediate steps).