system: remove page-dashboard-all, move remnants to page-login-logout; closes #7907
All the endpoints added here have no other reference and merely
exist as dashboard core additions.  Rename the visible ACL name
to hint at users that this is needed although it would be good to
offer this in the documentation for clarity.

Reusing this for maximum effect and least impact since it already
tried to create a dashboard minimal setup and here we also add the
license page and scrub the extra privilege as it does not seem to
be high profile and required.