MVC - move explicit cast in BaseModel when calling field->setValue() to offer the posibility to read structures in custom field types.
This is a requirement for https://github.com/opnsense/core/issues/7904

To avoid moving config items around, we need some way to handle <priv/> and <apikeys/> for example. When `setValue()` receives the `SimpleXMLElement`, we can parse the data into a flattened structure on initial read. All current callers either use an implicit string cast or an explicit one, only the BaseField and LinkAddressField types seem to need an explicit one.