System: Access - Restructure ACL privilege handling so we can simplify storage for https://github.com/opnsense/core/issues/7904
Currently both users and groups store multiple <priv/> tags within the same container, but if we want to use MVC structures, we need to move to a single container to hold the access roles for them.
Our common seperator is a comma (,) which is not used in acl tags as these are not valid within xml tags anyway, by supporting both the future storage privA,privB and the current one we can feed this change easily in a future minor release without having to move all new mvc code already.

After this change, a config.xml may contain (combinations of):

   <priv>page-X</priv>
   <priv>page-Y</priv>

   <priv>page-X,page-Y</priv>