firmware: generalise the output handling
This also moves the initial output to the start of the script
again which clears and seeds the output file so the GUI picks
up the right operation immediately.  This was a problem during
CRL setup which may delay the execution too much.

Make the output_done() and output_reboot() both never return
which simplifies the code a bit.

There is more potential for improved output handling but for
the time being this should suffice to fix the current glitches.