system: straighten out syslog-ng rc.d scripting
Make sure that 'pluginctl -s xxx stop' isn't used here as it
does not stop a disabled instance.  The point is moot for
syslog-ng since you cannot disable it but that only leaves
a tiny amount of daemons that still use it in core (nothing
in plugins uses it).

Also consolidate away from service(8) since we expect and
call the rc.d script directly.