Services: Unbound DNS - fix model migration, between versions the "domain" container was moved to "dots", which means we can only move items into the new spot during migrations of old configurations. closes https://github.com/opnsense/core/issues/8395
This is certainly a downside of data migrations in general, when looking at the actual target, we don't have all the versions in between available. which means breakage is possible when skipping a lot of versions on our end.

(cherry picked from commit 03577ba3ca185bad62b8e9e07c2ccb045508a1b2)