webui: "non root" user mode. closes https://github.com/opnsense/core/issues/8521
Allow the webui to run as wwwonly and move related (temp) files to their own directories so we limit the choice of mangling rights.

When trying to transition back from wwwonly to root, require a file (/var/run/www_non_root) to be removed via the console as an extra barrier.

When captive portal is used, the api dispatcher is forced to use wwwonly in all situations as the number of endpoints used is small and easy to validate (no legacy impact)