Firewall: Aliases - performance improvements when using large amounts of aliases, for https://github.com/opnsense/core/issues/8654
Extend basemodel with getNodeDescriptions() to ease dumping the data a user is presented into an array structure, which we can then use in ModelRelationField to cache
objects while persisted_at is equal to the previous time we needed the information.
The getCachedData() static method is added to BaseModel for easy reuse.