dnsmasq: fix edge cases writing dnsmasq-hosts
Since 0d566922 we expect a host name, but the domain is optional.
DHCP allows to omit the host name too so only write the host name
if it is there and only write the domain if it is there on top of
the host name.  Otherwise do not render anything.

(cherry picked from commit 3d91d080f42817515caadf93bf245351310536ca)