dnsmasq: fix edge cases writing dnsmasq-hosts
Since 4b394dfc we expect a host name, but the domain is optional.
DHCP allows to omit the host name too so only write the host name
if it is there and only write the domain if it is there on top of
the host name.  Otherwise do not render anything.