dnsmasq: Add checkbox to dialogHostOverride that can set domains as local (#8730)
Couple it with the dhcp-range domain dataset to ensure a unique dataset for all local domains.

This enables users to set domains as local in classic host overrides, if they want dnsmasq to handle this domain exclusively.

This should give flexibility for almost all DNS forwarding scenarios, even if DHCP is not used at all.